Setup github action notify when new SVN commits are available.
This could probably eventually be a straight run of the `migrate` script
and push to the repo, but from time to time I do have to adjust the
parent mapping or some other thing on the svn-all-fast-export config.

To at least automate things, this will compare the latest migrated SVN
revision to the latest revision upstream, then open a PR to track that
commit. I will then action it be migrating and pushing and then closing
the PR.